feat: add Tabs component with Radix UI integration and update dependencies
