+++ client/tsconfig.json
@@ -28,5 +28,5 @@
   },
   "include": [
     "src"
-  ]
+, "src/components/analysisModal", "src/components/Evaluations.tsx", "src/components/StudentForm.tsx", "src/components/StudentList.tsx", "src/components/Classes.tsx"  ]
 }
